requirements.txt
pyroblack
tgcrypto
python-dotenv
motor
redis
requests
Pillow
yt-dlp